pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for Dentist's appointment"
authors = ["Dmytro Bukrieiev <dbucoder@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
pydantic = {extras = ["email"], version = "^2.7.0"}
aiosqlite = "^0.20.0"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
mypy = "^1.9.0"
ruff = "^0.4.1"


[tool.black]
line-length = 88

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "UP", "I", "PL", "ISC", "RUF"]
unfixable = ["B"]
ignore = ["E501", "RUF001", "B024", "PLR2004", "RUF009", "PLR0913", "E402"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
pretty = false
check_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
